Use next-gen components for notebook
