+++ packages/core/src/providers/index.js
@@ -3,6 +3,5 @@ import CellCreator from "./cell-creator";
 import Editor from "./editor";
 import NotebookApp from "./notebook-app";
 import Toolbar from "./toolbar";
-import Cell from "./cell";
 
-export { CellCreator, Editor, NotebookApp, Toolbar, Cell };
+export { CellCreator, Editor, NotebookApp, Toolbar };
